# listener: minor improvements to listeners