This PR does the following
- Moves the `buildListenerOpts` and `plugin.InputParams` out of loop to reduce the number of allocations
- Improve some comments
- Cleanup tests
    - Move ProxyV14 tests to main tests as they are default now (remove v14 suffix)
   - Rename the old tests to `WithDisabledSniffing` to indicate they test disabled sniffing behaviour