@istio-testing 2020-04-06
@ramaraochavali: The following tests **failed**, say `/retest` to rerun all failed tests:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
integ-distroless-k8s-tests_istio | a83c991518e3ded8b08a34bd218fa66bd2f09a5b | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/22745/integ-distroless-k8s-tests_istio/4074) | `/test integ-distroless-k8s-tests_istio`
integ-pilot-k8s-tests_istio | a83c991518e3ded8b08a34bd218fa66bd2f09a5b | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/22745/integ-pilot-k8s-tests_istio/10912) | `/test integ-pilot-k8s-tests_istio`
integ-security-k8s-tests_istio | a83c991518e3ded8b08a34bd218fa66bd2f09a5b | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/22745/integ-security-k8s-tests_istio/11068) | `/test integ-security-k8s-tests_istio`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->