@ramaraochavali 2020-04-06
Benchmark results before/after allocation change
Before :
```
BenchmarkListenerGeneration/empty-12         	     134	   9295216 ns/op	 2773969 B/op	   37808 allocs/op
BenchmarkListenerGeneration/telemetry-12     	     100	  10570073 ns/op	 3170342 B/op	   42789 allocs/op
BenchmarkListenerGeneration/virtualservice-12        124	   9174470 ns/op	 2774603 B/op	   37812 allocs/op
```
After : 
```
BenchmarkListenerGeneration/empty-12         	     154	   7735643 ns/op	 2735749 B/op	   37410 allocs/op
BenchmarkListenerGeneration/telemetry-12     	     128	   9000032 ns/op	 3132330 B/op	   42391 allocs/op
BenchmarkListenerGeneration/virtualservice-12        154	   7369656 ns/op	 2736207 B/op	   37413 allocs/op
```